--- a/scripts/perf-fast.sh
+++ b/scripts/perf-fast.sh
@@ -1,4 +1,8 @@
 
+#!/bin/bash
+
 zig build build-evm-runner -Doptimize=ReleaseFast \
   && zig build build-orchestrator -Doptimize=ReleaseFast \
-  && ./zig-out/bin/orchestrator --compare --export markdown --num-runs 1 --js-runs 1 --internal-runs 10 --js-internal-runs 1
+  && ./zig-out/bin/orchestrator --compare --export markdown --num-runs 1 --js-runs 1 --internal-runs 10 --js-internal-runs 1 \
+  && echo "Opening results in browser..." \
+  && npx markserve bench/official/results.md