## Summary

This PR improves benchmark reporting by normalizing all timing results to per-run values and automatically selecting the most appropriate time unit (microseconds, milliseconds, or seconds).

### Key Changes

- **Per-run normalization**: All benchmark results now show time per individual execution run rather than total time for all internal runs
- **Smart unit selection**: Times automatically display in the most appropriate unit (μs, ms, or s) based on magnitude
- **Enhanced reporting**: Both console output and markdown exports now show normalized results with proper units
- **Internal run tracking**: Added `internal_runs` field to track how many runs were used for normalization
- **Clear documentation**: Updated all reports to explain the per-run normalization

### Technical Details

- Added `internal_runs: u32` field to `BenchmarkResult` struct
- Implemented normalization logic: `total_time / internal_runs`
- Created `FormattedTime` struct with smart unit selection and formatting
- Updated both single-EVM and comparison markdown exports
- Enhanced console output to show units and internal run counts

### Benefits

- **Fair comparisons**: Results are now comparable across different EVM implementations regardless of their internal execution patterns
- **Better readability**: Automatic unit selection makes results easier to read (e.g., "12.34 μs" instead of "0.01234 ms")
- **Clear transparency**: Shows exactly how many internal runs were used for each benchmark
- **Consistent reporting**: All output formats (console, markdown) use the same normalized approach

### Test Plan

- [x] Build and compile successfully
- [x] All existing tests pass
- [x] Verified normalization logic works correctly
- [x] Checked markdown generation produces proper formatting
- [x] Console output displays units correctly

🤖 Generated with [Claude Code](https://claude.ai/code)