@roninjin10 2025-08-05
## Performance Benchmark Results

**Commit**: 03f734b (🎉 feat: normalize benchmark results per internal run with smart unit display)

### Results (normalized per-run):
| Test Case | Guillotine | REVM |
|-----------|------------|------|
| erc20-approval-transfer | 5.95 ms | 5.35 ms |
| erc20-mint | 4.54 ms | 4.13 ms |
| erc20-transfer | 7.26 ms | 6.77 ms |
| ten-thousand-hashes | 2.79 ms | 1.79 ms |
| snailtracer | 210.02 ms | 35.96 ms |

### Performance Comparison to PR #352:
**DRAMATIC IMPROVEMENT! 🎉** This PR shows a return to FAST performance state:
- erc20-approval-transfer: **21.1x faster** (125.55 ms → 5.95 ms)
- erc20-mint: **20.8x faster** (94.28 ms → 4.54 ms)
- erc20-transfer: **20.8x faster** (150.72 ms → 7.26 ms)
- ten-thousand-hashes: **21.4x faster** (59.61 ms → 2.79 ms)
- snailtracer: **20.1x faster** (4.23 s → 210.02 ms)

### Notes:
- This PR introduces normalized per-run results with smart unit display
- Guillotine returns to competitive performance levels (within same order of magnitude as REVM)
- Results now show microsecond precision for better granularity
- All benchmarks run with 2 runs, 20 internal runs per benchmark case